Added configuration, simplifed css, image rendering
I firstly needed to change the image rendering in fluid so that the
images are not assigned width/height. Next I needed to remove the title
and description box so I added new configuration options in the
flexform. Last I modified the CSS to work with my YAML CSS framework.
Mostly this was the margins and padding assigned to UL's.